feat(ot): Wave 6 catalog promote vehicle + building lab codecs
Fix SOME/IP hex service/method parsing and BACnet MS/TP write detection;
catalog canopen/uds/someip/hart-ip/bacnet-mstp/wmbus (matrix 130, stubs 32).

+++ packages/ispf-driver-uds/src/test/java/com/ispf/driver/uds/UdsDeviceDriverTest.java
@@ -257,7 +257,7 @@ class UdsDeviceDriverTest {
             int sid = uds[0] & 0xFF;
             if (sid == UdsDeviceDriver.SID_DIAGNOSTIC_SESSION_CONTROL) {
                 sessionActive = true;
-                return new byte[]{(byte) 0x50, uds[1], 0x00, 0x32, 0x01, 0xF4};
+                return new byte[]{(byte) 0x50, uds[1], 0x00, 0x32, 0x01, (byte) 0xF4};
             }
             if (sid == UdsDeviceDriver.SID_READ_DATA_BY_IDENTIFIER) {
                 int did = ((uds[1] & 0xFF) << 8) | (uds[2] & 0xFF);
